refactor(core): use gql.tada on get web pages queries
