--- apps/core/client/fragments/web-page.ts
@@ -1,9 +1,0 @@
-export const WEB_PAGE_FRAGMENT = /* GraphQL */ `
-  fragment WebPage on WebPage {
-    __typename
-    entityId
-    parentEntityId
-    name
-    isVisibleInNavigation
-  }
-`;
