+++ apps/core/client/queries/get-web-page.ts
@@ -1,65 +1,52 @@
 import { cache } from 'react';
 
 import { client } from '..';
-import { graphql } from '../generated';
+import { graphql } from '../graphql';
 import { revalidate } from '../revalidate-target';
 
-export const GET_WEB_PAGE_QUERY = /* GraphQL */ `
-  query getWebPage($path: String!, $characterLimit: Int = 120) {
-    site {
-      route(path: $path) {
-        node {
-          ... on ContactPage {
-            contactFields
-            path
-            htmlBody
-            plainTextSummary(characterLimit: $characterLimit)
-            renderedRegions {
-              regions {
-                name
-                html
-              }
-            }
-            seo {
-              pageTitle
-              metaKeywords
-              metaDescription
-            }
+const WEB_PAGE_FRAGMENT = graphql(`
+  fragment WebPage on WebPage {
+    __typename
+    entityId
+    name
+    seo {
+      pageTitle
+      metaKeywords
+      metaDescription
+    }
+  }
+`);
+
+const GET_WEB_PAGE_QUERY = graphql(
+  `
+    query getWebPage($path: String!) {
+      site {
+        route(path: $path) {
+          node {
             ...WebPage
-          }
-          ... on NormalPage {
-            htmlBody
-            plainTextSummary(characterLimit: $characterLimit)
-            renderedRegions {
-              regions {
-                name
-                html
-              }
+            ... on ContactPage {
+              contactFields
+              htmlBody
             }
-            seo {
-              pageTitle
-              metaKeywords
-              metaDescription
+            ... on NormalPage {
+              htmlBody
             }
-            ...WebPage
           }
         }
       }
     }
-  }
-`;
+  `,
+  [WEB_PAGE_FRAGMENT],
+);
 
 export interface Options {
   path: string;
-  characterLimit?: number;
 }
 
-export const getWebPage = cache(async ({ path, characterLimit = 120 }: Options) => {
-  const query = graphql(GET_WEB_PAGE_QUERY);
-
+export const getWebPage = cache(async ({ path }: Options) => {
   const response = await client.fetch({
-    document: query,
-    variables: { path, characterLimit },
+    document: GET_WEB_PAGE_QUERY,
+    variables: { path },
     fetchOptions: { next: { revalidate } },
   });
